Add permissions and roles
I added permissions and roles. You must now be a 'blog-creator' to
create an post. You can register to become a user but that does not make
you a blog creator. Only blog creators can create new blog creators.

The route for adding a new blog creator is /blog-creator/create. This
will take you to a view where you can add a new blog creator by email
address.

I have set it up so that if the user already exists it merely makes that
user a blog creator.

However, if the user does not exist, I create a new user with a default
password.

Ideally, when that new user logs in they would be prompted to change
their password and name. I have not implemented that yet.

Some notable changes include the routing:

I removed the home controller.

The home controller was confusing. I thought it would be better to make
the home controller the articles index method--just show all of the
articles.